Deprecate menu code
